+++ tests/testthat/test-from_url.R
@@ -67,7 +67,7 @@ test_that("nflverse_download downloads files",{
       nflverse_download(
         "asdf",
         folder_path = temp_dir,
-        .token = Sys.getenv("NFLVERSE_GH_TOKEN")),
+        .token = Sys.getenv("NFLVERSE_GH_TOKEN", unset =  gh::gh_token())),
       "Could not find"
     ),
     "No matching releases"
@@ -78,7 +78,7 @@ test_that("nflverse_download downloads files",{
       combine, "test",
       folder_path = temp_dir,
       file_type = "qs",
-      .token = Sys.getenv("NFLVERSE_GH_TOKEN")
+      .token = Sys.getenv("NFLVERSE_GH_TOKEN", unset =  gh::gh_token())
     ),
     regexp = "Could not find file"
   )
@@ -87,7 +87,7 @@ test_that("nflverse_download downloads files",{
     combine, "contracts",
     folder_path = temp_dir,
     file_type = "parquet",
-    .token = Sys.getenv("NFLVERSE_GH_TOKEN")
+    .token = Sys.getenv("NFLVERSE_GH_TOKEN", unset =  gh::gh_token())
   )
 
   expect_true(
@@ -101,7 +101,7 @@ test_that("nflverse_releases lists releases",{
   skip_on_cran()
   skip_if_offline("github.com")
 
-  releases <- nflverse_releases(.token = Sys.getenv("NFLVERSE_GH_TOKEN"))
+  releases <- nflverse_releases(.token = Sys.getenv("NFLVERSE_GH_TOKEN", unset =  gh::gh_token()))
 
   expect_true(is.data.frame(releases))
   expect_true(nrow(releases) >= 15)