+++ tests/testthat/test-utils_sitrep.R
@@ -1,6 +1,20 @@
 test_that("sitrep works", {
   skip_on_cran()
-  expect_output(.sitrep("nflreadr", recursive = FALSE), regexp = "System Info|R version |Running under|nflreadr \\(")
-  expect_null(.sitrep("xyz", recursive = FALSE))
-  expect_output(nflverse_sitrep(pkg = "nflreadr"), regexp = "System Info|R version |Running under|nflverse Packages|nflreadr \\(|nflverse Dependencies")
+  nflreadr_sitrep <- .sitrep("nflreadr", recursive = FALSE)
+  expect_named(
+    nflreadr_sitrep,
+    c("system_info", "installed", "dependencies", "package_options",
+      "not_installed", "packages", "timestamp")
+  )
+
+  xyz_sitrep <- .sitrep("xyz", recursive = FALSE)
+
+  expect_true(
+    all(lengths(xyz_sitrep[c("installed", "dependencies", "package_options")]) == 0)
+  )
+
+  expect_output(
+    print(nflreadr_sitrep),
+    regexp = "System Info|R version |Running under|Packages|nflreadr \\("
+  )
 })